+++ test/dialect/mssql/test_compiler.py
@@ -20,6 +20,7 @@ from sqlalchemy import String
 from sqlalchemy import Table
 from sqlalchemy import testing
 from sqlalchemy import text
+from sqlalchemy import try_cast
 from sqlalchemy import union
 from sqlalchemy import UniqueConstraint
 from sqlalchemy import update
@@ -1477,10 +1478,15 @@ class CompileTest(fixtures.TestBase, AssertsCompiledSQL):
         metadata = MetaData()
         t1 = Table("t1", metadata, Column("id", Integer, primary_key=True))
 
-        self.assert_compile(
-            select(try_cast(t1.c.id, Integer)),
-            "SELECT TRY_CAST (t1.id AS INTEGER) AS id FROM t1",
-        )
+        def call(func):
+            self.assert_compile(
+                select(func(t1.c.id, Integer)),
+                "SELECT TRY_CAST (t1.id AS INTEGER) AS id FROM t1",
+            )
+
+        with testing.expect_deprecated(".*try_cast.*"):
+            call(mssql_base.try_cast)
+        call(try_cast)
 
     @testing.combinations(
         ("no_persisted", "", "ignore"),
